Fix SL6 Build Configure Race
gst_plugins_bad is an automake project, but comes with a pre-generated
configure file.

When building for SL6 using 'build-with-progress.sh', gst_plugins_bad is
first configured using the pre-generated configure file provided with the
package, prior to applying its patches and generating a new configure file.

This means that, where patches are attempting to make changes to
configure.ac, those patches are applied too late, because the project has
already been configured using the pre-generated configure script provided
in the tarball.

This fix pre-generates the configure script such that it is equivalent to
the script which the build would have generated, had it run autogen after
applying the patches, such that if it configures before running autogen,
then the configuration is correct.