fix deadlock if downloader is cancelled
The function gst_uri_downloader_fetch_uri_with_range deadlocks if
the uridownloader is cancelled. This is because it trys to call
gst_uri_downloader_stop while holding the object lock. This fails
because gst_uri_downloader_stop needs to change the element's state
and this change requires the object lock.