Implement live seek
hlsdemux assumes that seeking is not allowed for live streams,
however seek is possible if there are sufficient fragments in the
manifest. For example the BBC have live streams that contain 2 hours
of fragments.

This commit implements position query for live streams, to find the
available seek range and implements the seek event for live streams.