fix int overflow in GST_EVENT_SEEK
The seek code for both live and on-demand is common code. The
difference between them is that an offset has to be calculated
for the timecode of the first fragment in the live playlist.

This offset was being stored in a 32 bit integer which overflows
because gstreamer uses nanoseconds to represent time.