DEVARCH-6858: redesign of live seek support
The current implementation of live seek has a defect that causes
it to somtimes fail complete a seek to the end of a live stream.
This was due to an error in how hlsdemux produced the timestamps for
buffers following a seek.

When hlsdemux starts to play a live stream, the possible seek range
is between 0 and A seconds. After some time has passed, the beginning of
the stream will no longer be available in the playlist and the seek
range is between B and C seconds.

Seek range:
start          0 ........... A
later               B ........... C

If a seek is performed at any time when B>0, the code in hlsdemux that
sets the timestamp of the buffer was incorrect.

For example, if B=20,  C=120, seek=60:
hlsdemux would correctly select the fragment that contains time=60,
send a segment event for 00:00:60 -> 99:99:99, but then start
timestamping buffers starting at 40 (because 60-B=40).

This commit uses a new logic where it keeps a note of the B and C values
and the highest sequence number it has seen. Every time it updates the
media playlist, it walks the list of fragments, seeing if there is a
fragment with sequence number > highest_seen_sequence. If so, the values
of B and C are updated. The value of B is used when timestamping
buffers.