more fixes for shutdown race conditions in hlsdemux
During shutdown there are race conditions that can cause a segfault from
hlsdemux. This is because the client->current pointer becomes NULL
during the shutdown phase, once the request for a playlist update has
been cancelled. If the update task fires during this time, it causes a
null pointer dereference.