hlsdemux selects incorrect starting position for live streams
When playing live HLS streams, the media playback starts from the
beginning of the media playlist. When playing a live HLS stream, media
playback should start from 3 fragments from the end of the playlist.

See section 6.3.3. of the HLS draft [1]

  [1] http://tools.ietf.org/html/draft-pantos-http-live-streaming-12#page-29

This patch changes the logic to select 3 fragments from the end when
playing a live stream.