hlsdemux: handle case of overlapping seek events
If a seek event arrives while a seek is in progress, it is possible to
get in to a state where the active fragments are unref'd. This occurs
if a second seek arrives while a first seek is waiting for the hlsdemux
tasks to stop and for the buffers to be flushed. This second seek will
pass through this stop+flush phase very quickly and then start the
tasks to download the next fragments. When the first seek has finished
the flush phase, it does an unref of every fragment in hlsdemux.

This commit changes the logic in to three phases. A position
calculation, a stop+flush and a start. The "position calculation" is
protected by the client lock. This code was already 75% protected by the
client lock, it just needed a few lock-free versions of existing
functions.

The flushing phase is not mutex protected, as it needs to call functions
that require the client lock. Instead it uses the logic that if another
thread performs a position calculation while the flushing phase is in
progress, this second thread calculates the new position and then returns.
The first thread will finish its flush and then use this newly calculated
position when moving in to the starting phase. This is because the final
destination for the seek is now held in a variable in the demux class,
so that the last thread to send the seek event is the one that defines
the final seek position. This matches the desired behaviour for HTML5
media element.

The start phase is protected by the client lock, which was the existing
behaviour in hlsdemux.

This final part of the puzzle is if a seek arrives after the starting
phase has completed, but media playback hasn't started. In this case,
the downloads are cancelled and the stop+flush phase has to be
performed.