round seek position down to previous fragment boundary
Following a seek, the position reporting to WebKit is incorrect. The
position reported via the timeupdate event can be up to 10 seconds out
from what is being displayed on screen.

The issue is that hlsdemux seeks to a multiple of the fragment duration
(10 seconds in this case) and expects a gstreamer element downstream of
this to drop decoded frames between the actual seek point and the
fragment start.

Polonium Nexus sink assumes that the first frame it receives after a
segment event starts at the beginning of the segment.

This commit makes hlsdemux round down seek requests to fragment start,
regardless of the requested seek position.