hlsdemux: Fix hlsdemux abr if bitrate drops.
hlsdemux makes an incorrect assumption about how to calculate
download bitrate. It tries to download a fragment every target_duration
seconds and then calculates bitrate as:

    g_get_current_time (&now);
    diff = (GST_TIMEVAL_TO_TIME (now) -
        GST_TIMEVAL_TO_TIME (demux->next_update));

The problem with this code is that if hlsdemux ever fails to start a
download on time, the above calculation goes wrong for all subsequent
downloads.  Another problem with hlsdemux is that if the queue becomes
empty, it will never perform another download bitrate calculation.

commit 5d5ca5139e74cffd7c893d1fddabd0ec158d0804 from Sebastian Dröge
<sebastian@centricular.com> modifies this calculation to use the
start and stop times from uridownloader to calculate download bitrate.

If the queue becomes empty, as soon as a fragment is downloaded, it is
consumed by the stream task and sent downstream. The
gst_hls_demux_switch_playlist tries to peek at this queue to find the
last fragment, and then computes how long it took to download. The fix
for this problem is to return a reference to the fragment in
gst_hls_demux_get_next_fragment and unref this once the bitrate
calculation has been performed.