hlsdemux: Short circuit caching first fragments if link is fast enough.
If the download speed is more than twice the bandwidth required for
the currently selected variant, do not wait 3 fragments before starting
playback.

This patch is only usable if the update task is modified to allow
fragments to be downloaded faster than realtime. If the update task is
configured to allow downloads at 4x realtime, and the download speed is
2x realtime, then the total amount of time required to get the next
fragment will be:
    target_duration/2 + target_duration/4

which is less than target_duration, which is the amount of time it will
take the decoder to decode+display the previous fragment.