hlsdemux: Disconnect playlist update from fragment download.
After the initial start-up phase, hlsdemux will only download a fragment
every target_duration seconds. This means that if playback ever falls
behind, it can never catch up. It can also have problems if the actual
duration of each fragment is slightly less than the target_duration.

This commit creates a new next_playlist_update variable that is used to
record when to re-download a live playlist. The existing next_update is
kept, which causes the update task to fire at regular intervals. The
interval is reduced to 0.25 * target_duration under normal download
conditions. This allows fragments to be downloaded up to 4x real time.