hlsdemux: Only set end_of_playlist on live streams.
If the download queue becomes empty on a live stream, the
end_of_playlist flag is set to TRUE, causing gst_hls_demux_updates_loop
to send an end of stream (EOS) event, which causes playback to stop.

This commit changes this to only send an EOS for a VOD stream and to
reset the end_of_playlist for a live stream if a download is successful.