hlsdemux: Stop spamming a server that has a live playlist without a TARGETDURATION field.
If the live playlist does not contain a #EXT-X-TARGETDURATION field, the
gst_hls_demux_updates_loop task will spin in a tight loop because the
targetduration field is zero. Each spin causes a download of the playlist.

All HLS live playlists should have the #EXT-X-TARGETDURATION field in
media playlists, so this commit is defensive behaviour to stop the
client from DOS attacking a server with an invalid playlist.