hlsdemux: Switch bitrate during fast-start if bitrate is sufficient.
At the moment, if the quick-start algorithm is active, two fragments
are downloaded at the lowest bitrate before rate adaptation is
performed. On a fast Internet connection, this means that the user sees
20 seconds of lower quality video, when they could have only had to
put up with 10 seconds of lower quality video.

If the bitrate is sufficient to download the highest bitrate at least
2x realtime, it is safe to allow ABR during fast start.

This commit adds a check to find out what the next bitrate would be if
ABR were allowed, and then checking to see if the last fragment was
downloaded at more than twice this value. If so, ABR is enabled.

This causes a switch after the first fragment, if the bitrate is
sufficient.