hlsdemux: Fix intermittent segfault when downloading a fragment fails.
This fixes a segfault that can occur when a download is cancelled, due
to a null point dereference.

`GST_BUFFER_DURATION(buf)` and `GST_BUFFER_PTS(buf)` can segfault if a
download doesn't contain a buffer. This can happen if the download is
cancelled, which can happen during teardown or when switching bitrates.