hlsdemux: Avoid deadlock on element dispose by ensuring proper locking order
An erroneous locking order in gst_hls_demux_dispose would
cause a deadlock.

This is because mutex associated with a conditional variable
used to wake up the update thread would not be unlocked before
attempting to lock another mutex. This would cause update thread
to never wake up.

A fix removes an unnecessary lock/unlock pair in gst_hls_demux_dispose
to avoid the issue.